% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method for linreg objects}
\usage{
\method{coef}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named vector of coefficients.
}
\description{
This method returns the coefficients from a linreg object.
}
