\name{Four-package}
\alias{Four-package}
\alias{Four}
\docType{package}
\title{
\packageTitle{Four}
}
\description{
\packageDescription{Four}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Four}

Maintainer: \packageMaintainer{Four}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Four)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Four}}
%\section{Documentation Index}{\packageIndices{Four}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
