% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Function}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted (e.g., y ~ x1 + x2).}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class 'linreg' containing the following statistics:
  \item{coefficients}{Estimated coefficients for the regression model.}
  \item{fitted_values}{Fitted values from the regression.}
  \item{residuals}{Residuals (difference between observed and fitted values).}
  \item{degrees_of_freedom}{Degrees of freedom for the regression.}
  \item{residual_variance}{Estimated variance of the residuals.}
  \item{variance_coefficients}{Variance of the regression coefficients.}
  \item{t_values}{t-values for each coefficient.}
  \item{p_values}{p-values for each coefficient.}
}
\description{
This function performs multiple linear regression using ordinary least squares
and returns an object of class 'linreg' containing various statistics.
}
