% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object.}

\item{...}{Additional arguments (not used).}
}
\description{
This method plots the diagnostic plots for linreg objects.
}
