% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predict method for linreg class}
\usage{
\method{pred}{linreg}(x, ...)
}
\arguments{
\item{...}{Additional arguments}

\item{object}{An object of class linreg}
}
\description{
Predict method for linreg class
}
