% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg objects}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object.}

\item{...}{Additional arguments (not used).}
}
\description{
This method prints the coefficients and their names for linreg objects.
}
