% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{qr_decomposition}
\alias{qr_decomposition}
\title{QR Regression Function}
\usage{
qr_decomposition(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model to be fitted (e.g., y ~ x1 + x2).}

\item{data}{A data frame containing the variables in the model.}
}
\value{
An object of class 'qrreg' containing the following statistics:
  \item{coefficients}{Estimated coefficients for the regression model.}
  \item{variance_coefficients}{Variance of the regression coefficients.}
}
\description{
This function performs multiple linear regression using QR decomposition
and returns an object of class 'qrreg' containing various statistics.
}
