% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals method for linreg objects}
\usage{
\method{resid}{linreg}(x)
}
\arguments{
\item{x}{A linreg object.}
}
\value{
A vector of residuals.
}
\description{
This method returns the residuals from a linreg object.
}
