% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method for linreg objects}
\usage{
\method{summary}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object.}

\item{...}{Additional arguments (not used).}
}
\value{
A summary of the linreg object.
}
\description{
This method returns a summary of the linreg object, including coefficients,
standard errors, t-values, p-values, and estimate of σˆ.
}
