% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values method for linreg objects}
\usage{
\method{pred}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object.}

\item{...}{Additional arguments (not used).}
}
\value{
A vector of predicted values.
}
\description{
This method returns the predicted values from a linreg object.
}
